requirements.txt
pandas>=2.0
faker>=18.0
pyarrow>=12.0
apache-airflow>=2.7
